Use bool to ensure that numpy.bool is correctly compared.
Also, incorporate _lambda from modified PERT-distribution.
Next: Consider making the surjective constraint an argument
to the function instead of built-in. Also, consider allowing
specification of the correlation.  Granted, there are a limited
number of functions between 2 finite sets.  Thus, the desired
correlation may not be exactly attainable, but we should be
able to approximate it.